pyproject.toml
[tool.poetry]
name = "smolagentsfirstagent"
version = "0.1.0"
description = "Hugging Face agents course"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
smolagents = "^1.21.3"
ddgs = "^9.6.0"
opentelemetry-sdk = "^1.37.0"
opentelemetry-exporter-otlp = "^1.37.0"
openinference-instrumentation-smolagents = "^0.1.18"
langfuse = "^3.4.0"
litellm = "^1.77.1"
poetry-core = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
